refactor: move App component
